Defvar should define variable in the global environment
A following code in the emacs evaluates to 1:

(let ((x 2)) (defvar one 1))
one

This means that defvar defines variables in the global environment.

Also changed the way global environment is implemented. Implementation
with a static field is problematic for writing tests: since defvar and
defun modify global environment, we need a way to clear it between
tests. We can add method clear() to the environment, it will erase all
symbols in non-global environment and reinitialize symbols in global
environment. But I think a better way is to use non-static field,
because it makes a whole thing more flexible.